+++ backend/package.json
@@ -2,9 +2,9 @@
   "name": "backend",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
-    "serve": "node index.js",
+    "serve": "node src/index.js",
     "lint": "eslint --ext .js",
     "test": "jest"
   },
