+++ backend/src/posts/resolvers.js
@@ -1,0 +1,22 @@
+const { ApolloServer, gql } = require("apollo-server");
+
+const books = [
+    {
+        title: "Harry Potter and the Chamber of Secrets",
+        author: "J.K. Rowling"
+    },
+    {
+        title: "Jurassic Park",
+        author: "Michael Crichton"
+    },
+];
+
+const resolvers = {
+    Query: {
+      books: () => books,
+    },
+  };
+
+  module.exports = {
+    resolvers
+};
