+++ backend/src/posts/typeDefs.js
@@ -1,0 +1,17 @@
+const { gql } = require("apollo-server");
+
+const typeDefs = gql`
+
+  type Book {
+    title: String
+    author: String
+  }
+
+  type Query {
+    books: [Book]
+  }
+`;
+
+module.exports = {
+  typeDefs
+};
